Make column value extraction reusable, it's used in Oracle, Presto and Hive
