+++ presto/src/main/scala/com/yahoo/maha/executor/presto/PrestoQueryExecutor.scala
@@ -2,29 +2,17 @@
 // Licensed under the terms of the Apache License 2.0. Please see LICENSE file in project root for terms.
 package com.yahoo.maha.executor.presto
 
-import java.math.MathContext
 import java.sql.{ResultSet, ResultSetMetaData}
 
-import com.google.common.cache.{CacheBuilder, CacheLoader}
 import com.yahoo.maha.core._
-import com.yahoo.maha.core.query._
+import com.yahoo.maha.core.query.{ResultSetTransformer, _}
 import com.yahoo.maha.jdbc.JdbcConnection
 import grizzled.slf4j.Logging
-import org.joda.time.format.DateTimeFormat
-import org.joda.time.{DateTime, DateTimeZone}
 
-import scala.math.BigDecimal.RoundingMode
 import scala.util.{Failure, Try}
-
-import com.yahoo.maha.core.query.ResultSetTransformer$
 
 trait PrestoQueryTemplate {
   def buildFinalQuery(query: String,  queryContext: QueryContext, queryAttributes: QueryAttributes): String
-}
-
-object PrestoQueryExecutor {
-  final val DATE_TIME_FORMATTER = DateTimeFormat.forPattern("yyyy-MM-dd HH:mm:ss")
-  final val DATE_FORMATTER = DateTimeFormat.forPattern("yyyy-MM-dd")
 }
 
 class PrestoQueryExecutor(jdbcConnection: JdbcConnection,
@@ -33,92 +21,7 @@ class PrestoQueryExecutor(jdbcConnection: JdbcConnection,
                           transformers: List[ResultSetTransformer] = ResultSetTransformer.DEFAULT_TRANSFORMS) extends QueryExecutor with Logging {
   val engine: Engine = PrestoEngine
 
-  val mathContextCache = CacheBuilder
-    .newBuilder()
-    .maximumSize(100)
-    .concurrencyLevel(Runtime.getRuntime.availableProcessors())
-    .build(new CacheLoader[java.lang.Integer, MathContext]() {
-    override def load(key: java.lang.Integer): MathContext = {
-      new MathContext(key, java.math.RoundingMode.HALF_EVEN)
-    }
-  })
-  
-  def getBigDecimalSafely(resultSet: ResultSet, index: Int) : BigDecimal = {
-    Try[BigDecimal](resultSet.getBigDecimal(index)).getOrElse(0.0)
-  }
-  
-  def getLongSafely(resultSet: ResultSet, index: Int) : Long = {
-    val result = getBigDecimalSafely(resultSet, index)
-    Try[Long](result.longValue()).getOrElse(0L)
-  }
-
-  def getColumnValue(index: Int, column: Column, resultSet: ResultSet) : Any = {
-    column.dataType match {
-      case IntType(_, sm, _, _, _) =>
-        if(sm.isDefined) {
-          resultSet.getString(index)
-        } else {
-          getLongSafely(resultSet, index)
-        }
-      case DecType(len: Int, scale, _, _, _, _) =>
-        val result: java.lang.Double = {
-          if (scale > 0 && len > 0) {
-            val result = getBigDecimalSafely(resultSet, index)
-            if(result != null) {
-              val mc = mathContextCache.get(len)
-              result.setScale(scale, RoundingMode.HALF_EVEN).round(mc).toDouble
-            } else {
-              null
-            }
-          } else if(scale > 0) {
-            val result = getBigDecimalSafely(resultSet, index)
-            if(result != null) {
-              result.setScale(scale, RoundingMode.HALF_EVEN).toDouble
-            } else {
-              null
-            }
-          } else if(len > 0) {
-            val result = getBigDecimalSafely(resultSet, index)
-            if(result != null) {
-              val mc = mathContextCache.get(len)
-              result.setScale(10, RoundingMode.HALF_EVEN).round(mc).toDouble
-            } else {
-              null
-            }
-          } else {
-            val result = getBigDecimalSafely(resultSet, index)
-            if(result != null) {
-              result.setScale(10, RoundingMode.HALF_EVEN).toDouble
-            } else {
-              null
-            }
-          }
-        }
-        result
-      case dt: DateType =>
-        if(dt.format.isDefined) {
-          resultSet.getString(index)
-        } else {
-          val date = resultSet.getDate(index)
-          if (date != null) {
-              new DateTime(date).toString(PrestoQueryExecutor.DATE_FORMATTER)
-          } else {
-            null
-          }
-        }
-      case tt: TimestampType =>
-        val ts = resultSet.getTimestamp(index)
-        if (ts != null) {
-          val dateTime = new DateTime(ts, DateTimeZone.UTC)
-          dateTime.toString(PrestoQueryExecutor.DATE_TIME_FORMATTER)
-        } else {
-          null
-        }
-      case s: StrType =>
-        resultSet.getString(index)
-      case any => throw new UnsupportedOperationException(s"Unhandled data type for column value extraction : $any")
-    }
-  }
+  val columnValueExtractor = new ColumnValueExtractor
 
   def execute[T <: RowList](query: Query, rowList: T, queryAttributes: QueryAttributes) : QueryResult[T] = {
     val acquiredQueryAttributes = lifecycleListener.acquired(query, queryAttributes)
@@ -164,7 +67,7 @@ class PrestoQueryExecutor(jdbcConnection: JdbcConnection,
                   (alias, column) <- aliasColumnMap
                 } {
                   val index = columnIndexMap(alias)
-                  val value = getColumnValue(index, column, resultSet)
+                  val value = columnValueExtractor.getColumnValue(index, column, resultSet)
                   var transformedValue = value
                   transformers.foreach(transformer => {
                     val grain = query.queryContext.requestModel.queryGrain.getOrElse(DailyGrain)
